# Bypass tracing handler if tracing is disabled