<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

This drops a whopping 55 allocations **per request** from both the activator and the queue-proxy if tracing is disabled. We'd love to reduce the tracing overhead in general of course, but this at least relieves the situation for people that are not running tracing at all.

Benchmark results

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkHandlerChain/sequential-16     21598         13875         -35.76%
BenchmarkHandlerChain/parallel-16       20688         14301         -30.87%

benchmark                               old allocs     new allocs     delta
BenchmarkHandlerChain/sequential-16     125            70             -44.00%
BenchmarkHandlerChain/parallel-16       125            70             -44.00%

benchmark                               old bytes     new bytes     delta
BenchmarkHandlerChain/sequential-16     7701          4686          -39.15%
BenchmarkHandlerChain/parallel-16       7849          4793          -38.93%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @julz @vagababov